requirements.txt
certifi==2021.10.8
click==8.0.3
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
wincertstore==0.2
werkzeug-2.0.2
dataclasses-0.8
flask-2.0.2
click-8.0.3
markupsafe-2.0.1
itsdangerous-2.0.1
Jinja2>=3.0
numpy>=1.19.5
scipy>=1.5.4
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
pillow>=8.4.0
gunicorn==20.0.4
